Description
Poster submission for Phase 1 of Manhattan College's 2024 iteration of its renowned Business Analytics Competition (BAC@MC), submitted by myself on behalf of the David Nazarian College of Business & Economics as one of its elected representatives to the competition.
https://manhattan.edu/academics/schools-and-departments/school-of-business/business-analytics-competition/index.php

The 2024 Competition's dataset centered on U.S. nursing homes, containing financial, quality, health, penalty, COVID-19 vaccination, and provider information data from 37 different CSV files for the years 2015-2022 summing to tens of million of rows. 

Phase 1 of the competition had teams analyze the dataset (released in February 2024) in response to predetermined research objectives and prepare a poster for presentation at Manhattan College for May 20th. Phase 2, commencing the day after the end of Phase 1, presents teams with new data and asks them to perform additional analysis in response to new questions, culminating in a slide deck presentation of findings at the start of the following day.