Disclaimer
Due to the sheer volume and interconnectivity of code written for the Manhattan College's 2024 Business Analytics Competition, this sample has been limited to code unrelated to the competition. A sample of code related to any part of the 2024 Business Analytics Competition (visualizations, data cleaning/preprocessing, computations, etc.), which contains some of my most advanced work, can be produced upon request.